refactor: move layout components to src/components/layout
